refactor(sbb-icon-sidebar): extract to `icon-sidebar` module
