+++ src/elements/icon-sidebar.ts
@@ -1,0 +1,8 @@
+/**
+ * @entrypoint
+ */
+export * from './icon-sidebar/icon-sidebar.ts';
+export * from './icon-sidebar/icon-sidebar-button.ts';
+export * from './icon-sidebar/icon-sidebar-container.ts';
+export * from './icon-sidebar/icon-sidebar-content.ts';
+export * from './icon-sidebar/icon-sidebar-link.ts';
