+++ src/elements/icon-sidebar/icon-sidebar-button.ts
