+++ src/elements/icon-sidebar/icon-sidebar-button/icon-sidebar-button.component.ts
@@ -1,21 +1,21 @@
 import type { CSSResultGroup, TemplateResult } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { SbbLinkBaseElement } from '../../core/base-elements.js';
+import { SbbButtonBaseElement } from '../../core/base-elements.js';
 import { boxSizingStyles } from '../../core/styles.js';
 import { SbbIconNameMixin } from '../../icon.js';
+import { iconSidebarButtonCommonStyle } from '../../sidebar/common.js';
 import type { SbbTooltipDefaultPositions } from '../../tooltip.js';
-import { iconSidebarButtonCommonStyle } from '../common.js';
 
 /**
- * Link to be placed inside `sbb-icon-sidebar`.
+ * Button to be placed inside `sbb-icon-sidebar`.
  *
  * @slot icon - Slot used to display the icon.
  */
 export
-@customElement('sbb-icon-sidebar-link')
-class SbbIconSidebarLinkElement
-  extends SbbIconNameMixin(SbbLinkBaseElement)
+@customElement('sbb-icon-sidebar-button')
+class SbbIconSidebarButtonElement
+  extends SbbIconNameMixin(SbbButtonBaseElement)
   implements SbbTooltipDefaultPositions
 {
   public static override styles: CSSResultGroup = [boxSizingStyles, iconSidebarButtonCommonStyle];
@@ -30,6 +30,6 @@ class SbbIconSidebarLinkElement
 declare global {
   interface HTMLElementTagNameMap {
     // eslint-disable-next-line @typescript-eslint/naming-convention
-    'sbb-icon-sidebar-link': SbbIconSidebarLinkElement;
+    'sbb-icon-sidebar-button': SbbIconSidebarButtonElement;
   }
 }
