+++ src/elements/icon-sidebar/icon-sidebar-button/icon-sidebar-button.snapshot.spec.ts
