+++ src/elements/icon-sidebar/icon-sidebar-button/icon-sidebar-button.visual.spec.ts
