+++ src/elements/icon-sidebar/icon-sidebar-container.ts
