+++ src/elements/icon-sidebar/icon-sidebar-container/icon-sidebar-container.component.ts
@@ -1,7 +1,7 @@
 import { type CSSResultGroup, html, LitElement, type TemplateResult } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { sidebarContainerCommonStyle } from '../common.js';
+import { sidebarContainerCommonStyle } from '../../sidebar/common.js';
 import type { SbbIconSidebarElement } from '../icon-sidebar.js';
 
 import style from './icon-sidebar-container.scss?lit&inline';
