+++ src/elements/icon-sidebar/icon-sidebar-container/icon-sidebar-container.scss
