+++ src/elements/icon-sidebar/icon-sidebar-container/icon-sidebar-container.snapshot.spec.ts
