+++ src/elements/icon-sidebar/icon-sidebar-content.ts
