+++ src/elements/icon-sidebar/icon-sidebar-content/icon-sidebar-content.component.ts
@@ -1,7 +1,7 @@
 import { type CSSResultGroup, html, LitElement, type TemplateResult } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { sidebarContentCommonStyle } from '../common.js';
+import { sidebarContentCommonStyle } from '../../sidebar/common.js';
 
 import style from './icon-sidebar-content.scss?lit&inline';
