+++ src/elements/icon-sidebar/icon-sidebar-content/icon-sidebar-content.scss
