+++ src/elements/icon-sidebar/icon-sidebar-content/icon-sidebar-content.stories.ts
@@ -6,6 +6,8 @@ import readme from './readme.md?raw';
 
 import '../../card.js';
 
+import './icon-sidebar-content.component.js';
+
 const Template = (): TemplateResult => html`
   <sbb-card color="milk"> See 'sbb-icon-sidebar' examples to see it in action. </sbb-card>
 `;
@@ -20,7 +22,7 @@ const meta: Meta = {
       extractComponentDescription: () => readme,
     },
   },
-  title: 'elements/sbb-sidebar/sbb-icon-sidebar-container',
+  title: 'elements/sbb-icon-sidebar/sbb-icon-sidebar-content',
 };
 
 export default meta;
