+++ src/elements/icon-sidebar/icon-sidebar-link/icon-sidebar-link.component.ts
@@ -1,21 +1,21 @@
 import type { CSSResultGroup, TemplateResult } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { SbbButtonBaseElement } from '../../core/base-elements.js';
+import { SbbLinkBaseElement } from '../../core/base-elements.js';
 import { boxSizingStyles } from '../../core/styles.js';
 import { SbbIconNameMixin } from '../../icon.js';
+import { iconSidebarButtonCommonStyle } from '../../sidebar/common.js';
 import type { SbbTooltipDefaultPositions } from '../../tooltip.js';
-import { iconSidebarButtonCommonStyle } from '../common.js';
 
 /**
- * Button to be placed inside `sbb-icon-sidebar`.
+ * Link to be placed inside `sbb-icon-sidebar`.
  *
  * @slot icon - Slot used to display the icon.
  */
 export
-@customElement('sbb-icon-sidebar-button')
-class SbbIconSidebarButtonElement
-  extends SbbIconNameMixin(SbbButtonBaseElement)
+@customElement('sbb-icon-sidebar-link')
+class SbbIconSidebarLinkElement
+  extends SbbIconNameMixin(SbbLinkBaseElement)
   implements SbbTooltipDefaultPositions
 {
   public static override styles: CSSResultGroup = [boxSizingStyles, iconSidebarButtonCommonStyle];
@@ -30,6 +30,6 @@ class SbbIconSidebarButtonElement
 declare global {
   interface HTMLElementTagNameMap {
     // eslint-disable-next-line @typescript-eslint/naming-convention
-    'sbb-icon-sidebar-button': SbbIconSidebarButtonElement;
+    'sbb-icon-sidebar-link': SbbIconSidebarLinkElement;
   }
 }
