+++ src/elements/icon-sidebar/icon-sidebar-link/icon-sidebar-link.visual.spec.ts
