+++ src/elements/icon-sidebar/icon-sidebar.ts
