+++ src/elements/icon-sidebar/icon-sidebar/icon-sidebar.component.ts
