+++ src/elements/icon-sidebar/icon-sidebar/icon-sidebar.scss
