+++ src/elements/icon-sidebar/icon-sidebar/icon-sidebar.snapshot.spec.ts
