+++ src/elements/icon-sidebar/icon-sidebar/icon-sidebar.spec.ts
