+++ src/elements/sidebar.ts
@@ -2,11 +2,6 @@
  * @entrypoint
  */
 export * from './sidebar/common.js';
-export * from './sidebar/icon-sidebar.js';
-export * from './sidebar/icon-sidebar-button.js';
-export * from './sidebar/icon-sidebar-container.js';
-export * from './sidebar/icon-sidebar-content.js';
-export * from './sidebar/icon-sidebar-link.js';
 export * from './sidebar/sidebar.js';
 export * from './sidebar/sidebar-close-button.js';
 export * from './sidebar/sidebar-container.js';
